requirements.txt
beanie
fastapi
httpx
motor
pydantic
pydantic-settings
python-dotenv
uvicorn